+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Replaced `AuthToken.value` with `AuthToken.toString()`. 
 - Supported get `AsyncTextGeneration` Operation status.
 - Converted `YandexGptApi` into its implementation `YandexGptApiClient`.
+- Renamed `TextGenerationResponse` into `TextCompletion`.
 
 ## 0.4.0